backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.9
Pillow==9.5.0
PyYAML==6.0
python-dotenv==1.0.1
django-filter==23.5